requirements.txt
redis==5.0.1
rq==1.15.1
python-dotenv==1.0.0
supabase==2.3.0
httpx==0.24.1
yt-dlp==2026.02.04
ffmpeg-python==0.2.0
Pillow==12.1.0
numpy==1.26.3
torch==2.10.0
transformers==4.45.0
anthropic==0.39.0
youtube-transcript-api==1.2.4
openai-whisper==20250625
fastapi==0.115.0
uvicorn==0.32.0
pytest==8.3.4
